+++ R/rcpp.R
@@ -1,0 +1,3 @@
+#' @useDynLib nngeo
+#' @importFrom Rcpp sourceCpp
+NULL