+++ NEWS.md
@@ -1,6 +1,9 @@
-# nngeo 0.1.0
+# nngeo 0.1.0 (2018-01-16)
 
 * Initial version
 
+# nngeo 0.1.2
 
-
+* Added tests for 'st_nn'
+* Added progress bar for 'st_nn'
+* Added C code (Vincenty) for 'st_nn' & lon-lat points